input: synaptics_dsx_i2c: Use LCD notifier instead of FB notifier
Plugging up the PM ops directly to the FB notifier and blocking it is bad, and can cause severe lag when waking the device from sleep, as well as deadlock FB.

Use LGE's LCD notifier implementation instead. Crucial FB PM ops are now faster as a result (due to the lack of a blocker), so a framebuffer is generated after waking from sleep faster than before.

Signed-off-by: Sultanxda <sultanxda@gmail.com>