USB: dwc3_otg: Fix race with i2c-qup when entering from power collapse
When the device is power-collapsed and an OTG peripheral is inserted, DWC3
wakes the device and proceeds to enable host mode; however, it does this very
quickly. Our charger controller is connected to an I2C bus, so in order to
access it and enable host mode, the I2C bus must be active. DWC3 attempts to
enable host mode ~4ms before the I2C driver finishes resuming, causing the
attempt by DWC3 to fail spectacularly (OTG peripheral is not detected and all
normal USB operations do not work until the OTG cable is reinserted while the
I2C bus is active).

In order to prevent host mode from failing, stall DWC3 until the I2C bus our
charger (bq24196) is connected to finishes resuming.

Signed-off-by: Sultanxda <sultanxda@gmail.com>