msm_thermal: implement a few improvements
So recently, before finalizing the driver, Franco had been working on a few changes that he abandoned to improve it. I took what was useful from those changes and added them here. I should mention though, that a bit of the work is my own and not Franco's.

1. Cores 1-3 will start shutting off whenever the third threshold is met after the 5th and 4th. This really helps keep the device cool in intense situations, and it does not seem to cause any lag in the userspace. I have tested this with numerous benchmarks and games over the last few hours and it works flawlessly.
2. When checking the temperature, we call a basic worqueue struct named thermal_wq. It acts as a basic worqueue until later when it is assigned with a high priority... It is TOTALLY safe to use a power effecient worqueue here as the initial purpose of the thermal_wq is still served
3. We now use an arch initcall... This is a thermal driver so init related stuff should be brought up earlier
4. pr_info is now ratelimited to avoid spam
5. Just did a general cleanup too...

That's all. Working amazingly well for me, and I am quite happy with this solution now. I may make further adjustments in the future, but for now, this is great